requirements.txt
langchain
langchain-community
langchain_openai
tiktoken
faiss-cpu
pdfplumber
slack-sdk
python-dotenv